move random rotation to to augmentation.py
